# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todo API",
        "description": "This is an open and free to use API for developers to test their client projects.\n        All api calls require you to pass in <br> your API Key with the querystring\n        <b>?apikey=########-####-####-####-############</b>, <br><br>Please\n        register for your own API Key today at https://todoapi.net/register ",
        "version": "1.0"
    },
    "paths": {
        "/api/keys/register": {
            "post": {
                "tags": [
                    "Integrations"
                ],
                "summary": "Register for new api key",
                "operationId": "IntegrationController.Register",
                "requestBody": {
                    "description": "Pass email and captcha",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "captcha"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "captcha": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registration successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "We have emailed you your api key."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/keys/forgot": {
            "post": {
                "tags": [
                    "Integrations"
                ],
                "summary": "Resend apikey to email on file",
                "operationId": "IntegrationController.Forgot",
                "requestBody": {
                    "description": "Pass email and captcha",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "captcha"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "captcha": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registration successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "We have emailed you your api key again."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/keys/deregister": {
            "post": {
                "tags": [
                    "Integrations"
                ],
                "summary": "Delete your developer account",
                "operationId": "IntegrationController.Deregister",
                "requestBody": {
                    "description": "Pass email and code",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "23445"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Deactivation request received",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "If code not provided, we have emailed you code to complete account deletion."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}